requirements.txt
pytesseract
Pillow
PyPDF2
rdflib
requests
pdfplumber